Moved game class behind trait for better testing and separation

+++ src/scajong/model/Tile.scala
@@ -12,11 +12,11 @@ class TilePair(val tile1:Tile, val tile2:Tile)
 
 class Tile(val x:Int, val y:Int, val z:Int, var tileType:TileType) {
   
-  def isInside(xf:Float, yf:Float) : Boolean = {
+  def isInside(xf:Double, yf:Double) : Boolean = {
     xf >= x && yf >= y && xf < x + Tile.Width && yf < y + Tile.Height
   }
 
-  def isInside(xf:Float, yf:Float, zf:Float) : Boolean = {
+  def isInside(xf:Double, yf:Double, zf:Double) : Boolean = {
     isInside(xf, yf) && zf >= z && zf < z + Tile.Depth
   }
